Move subgraph state utils to a separate file
